Introduce ttl_now
The new variable ttl_now in struct ctx is used for all calculations
on the "ttl" variables obj.ttl and beresp.ttl. This is necessary to
harmonize calculations between C and VCL code.

The change in struct vrt is part of the change in the VRT API. The
existing variable ctx.now stays and keeps it role as "the latest
timestamp taken before this sub was called". VMOD authors should
choose wisely on when to use now or ttl_now.