Remove t_restart from struct req, use t_req as ctx.now
Necessary to make c00041.vtc pass is a change from ">=" to ">" in
HSH_Lookup, which feels very OK.

However, this commit may very well break vmods that do purge-like
actions on the cache, and then a restart to fetch a new object. For
this reason vmods must use VRT_purge, and not HSH_purge with
ctx.now as the "now" argument.